Extract navigation code to separate files + rename for consistency
This fixes https://github.com/TelemetryDeck/SwiftSDK/issues/162
